Clean-up_first - change from int to Long

+++ src/main/java/com/kodilla/ecommercee/controller/CartController.java
@@ -14,7 +14,7 @@ public class CartController {
     private final List<CartDto> carts = new ArrayList<>();
 
     @GetMapping("/{cartId}")
-    public List<String> getCartById(@PathVariable int cartId) {
+    public List<String> getCartById(@PathVariable Long cartId) {
         List<String> productsName = new ArrayList<>();
         productsName.add("XXX");
         productsName.add("YYY");
@@ -27,17 +27,17 @@ public class CartController {
     }
 
     @PutMapping("/add/{cartId}/{productId}")
-    public CartDto addProductToCart(@PathVariable int cartId, @PathVariable int productId) {
+    public CartDto addProductToCart(@PathVariable Long cartId, @PathVariable Long productId) {
         return new CartDto(1L,1L,new ArrayList<>());
     }
 
     @DeleteMapping("/delete/{cartId}/{productId}")
-    public void removeProductFromCart(@PathVariable int cartId, @PathVariable int productId) {
+    public void removeProductFromCart(@PathVariable Long cartId, @PathVariable Long productId) {
         System.out.println("Product no." + productId + " has been removed from the cart no. " + cartId);
     }
 
     @PostMapping("/order/{cartId}")
-    public OrderDto createOrderFromCart(@PathVariable int cartId) {
+    public OrderDto createOrderFromCart(@PathVariable Long cartId) {
         return new OrderDto(1L,"1",1L, new ArrayList<>());
     }
 }
